requirements.txt
beautifulsoup4==4.4.1
google-api-python-client==1.4.2
httplib2==0.9.2
oauth2client==1.5.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
rsa==3.3
simplejson==3.8.1
six==1.10.0
titlecase==0.8.1
uritemplate==0.6
wheel==0.24.0
